% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSlopeTestData.R
\name{createSlopeTestData}
\alias{createSlopeTestData}
\title{A helper function to create a dataframe of breakpoints and associated \emph{p}-values from a \code{marge} model.}
\usage{
createSlopeTestData(
  marge.model = NULL,
  pt = NULL,
  is.gee = FALSE,
  is.glmm = FALSE
)
}
\arguments{
\item{marge.model}{A \code{marge} model object, like those returned from \code{\link{marge2}}. Defaults to NULL.}

\item{pt}{A data.frame containing pseudotime or latent time values. Defaults to NULL.}

\item{is.gee}{Was the GEE mode used? Defaults to FALSE.}

\item{is.glmm}{Was the GLMM mode used? Defaults to FALSE.}
}
\value{
A data.frame containing model data.
}
\description{
Creates a data.frame of \code{marge} model breakpoints, \emph{p}-values, and other info.
}
\seealso{
\code{\link{marge2}}

\code{\link{testSlope}}
}
\author{
Jack R. Leary
}
