% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelList.R
\name{getModelList}
\alias{getModelList}
\title{Retrieve models from \code{testDynamic()} output.}
\usage{
getModelList(test.results = NULL, which.model = "marge")
}
\arguments{
\item{test.results}{The nested-list output of \code{testDynamic()}. Defaults to NULL.}

\item{which.model}{Whether the \code{marge} or null models should be returned. Defaults to "marge".}
}
\value{
A list of \code{marge} or null model NB GLM objects.
}
\description{
This function retrieves a list of either \code{marge} or null models from the output of \code{testDynamic()}.
}
\examples{
\dontrun{getModelList(test.results = gene_test_list)}
\dontrun{getModelList(test.results = gene_test_list, which.model = "null")}
}
\author{
Jack Leary
}
