% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marge2.R
\name{marge2}
\alias{marge2}
\title{\code{marge2} models for negative binomial data.}
\usage{
marge2(
  X_pred = NULL,
  Y = NULL,
  pen = 2,
  tols_score = 1e-05,
  M = 7,
  minspan = NULL
)
}
\arguments{
\item{X_pred}{: a matrix of the predictor variables. Note that this matrix should include a column of 1's for the intercept term.}

\item{Y}{: the response variable. A vector of length n by N.}

\item{pen}{: a set penalty used for the GCV (note: MARGE doesn't actually use this). The default is 2.}

\item{tols_score}{: the set tolerance for monitoring the convergence for the difference in score statistics between the parent and candidate model (this is the lack-of-fit criterion used for MARGE). The default is 0.00001}

\item{M}{: a set threshold for the number of basis functions to be used. The default is 7.}

\item{minspan}{: a set minimum span value. The default is \code{minspan = NULL}.}
}
\value{
\code{marge} returns a list of calculated values consisting of:

\code{B_final} : the basis model matrix for the final model fit.

\code{wic_mat} : a matrix of WIC values (with both penalties) for MARGE models given by the forward pass.

\code{min_wic} : the WIC (with both penalties) for the final MARGE model.

\code{GCV} : the GCV for the final selected model.

\code{y_pred} : the fitted values from the final selected model (with both penalties).

\code{final_mod} : the final selected (with both penalties) model matrix.
}
\description{
MARS fitting function for generalized linear models (GLM).
}
\details{
For further details please look at the \code{mars_ls} function - there are more details on the general MARS algorithm. MARGE will produce output for two penalties: 2 and log(N). A figure is automatically generated plotting WIC against the no. of parameters.
}
\examples{
marge(X_pred = pseudotime_df, Y = expr_vec)
}
\references{
Friedman, J. (1991). Multivariate adaptive regression splines. \emph{The Annals of Statistics}, \strong{19}, 1--67.

Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, \strong{70}, 110--120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, 245--253.
}
\seealso{
\code{\link{mars_ls}} and \code{\link{backward_sel_WIC}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
