% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLRT.R
\name{ModelLRT}
\alias{ModelLRT}
\title{Perform a likelihood ratio test for one model against another.}
\usage{
ModelLRT(mod.1 = NULL, mod.0 = NULL)
}
\arguments{
\item{mod.1}{The model corresponding to the alternative hypothesis. Defaults to NULL.}

\item{mod.0}{The model corresponding to the null hypothesis. Defaults to NULL.}
}
\value{
A list containing the LRT test statistic, degrees freedom, and the \emph{p}-value computed using the Chi-squared assumption.
}
\description{
This function compares two models using a likelihood ratio test(LRT) under the assumption that the test statistic is asymptotically Chi-squared with degrees freedom equal to the difference in the number of parameters between the larger and smaller model.
}
\examples{
ModelLRT(mod1 = saturated_glm, mod0 = null_glm)
}
