% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbGAM.R
\name{nbGAM}
\alias{nbGAM}
\title{Fit a negative-binomial GAM.}
\usage{
nbGAM(expr = NULL, pt = NULL, theta.init = NULL)
}
\arguments{
\item{expr}{A vector of integer counts. Defaults to NULL.}

\item{pt}{A dataframe or vector of pesudotime values. Defaults to NULL.}

\item{theta.init}{(Optional) If you have a numeric value for \eqn{\theta} you'd like to use provide it here, else it can be estimated internally. Defaults to NULL.}
}
\value{
An object of class \code{gam}
}
\description{
Fits a negative-binomial family GAM (glorified wrapper function).
}
\examples{
nbGAM(expr_vec, pt_df)
nbGAM(expr_vec, pt_vec, theta.init = 2.5)
}
