% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMARGE.R
\name{PlotMARGE}
\alias{PlotMARGE}
\title{Plot results of \code{marge} models using \code{ggplot2}.}
\usage{
PlotMARGE(
  model = NULL,
  gene.counts = NULL,
  pt = NULL,
  gene = NULL,
  null.mod = NULL
)
}
\arguments{
\item{model}{The \code{marge} model to extract fitted values from. Defaults to NULL.}

\item{gene.counts}{A vector of integer expression values for each cell. Defaults to NULL.}

\item{pt}{A data.frame of pseudotime values for each cell. Defaults to NULL.}

\item{gene}{The name of the gene that's being analyzed. Used as the title of the \code{ggplot} object. Defaults to NULL.}

\item{null.mod}{(Optional) An \code{lm} or \code{glm} object specifying the intercept-only null model for comparison. Defaults to NULL.}
}
\value{
A \code{ggplot} object.
}
\description{
This function visualizes the fitted values of a \code{marge} model over the expression and pseudotime values of each cell.
}
\examples{
PlotMARGE(model = marge_mod, gene.counts = exp_vec, pt = pt_df, gene = "BRCA2", null.mod = lm(exp_vec ~ 1))
}
