% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scLANE_models}
\alias{scLANE_models}
\title{An object of class \code{scLANE}.}
\format{
An object of class \code{scLANE}.
}
\usage{
data(scLANE_models)
}
\description{
Contains the results from running \code{\link{testDynamic}} on \code{\link{sim_counts}}.
}
\keyword{datasets}
