% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_counts}
\alias{sim_counts}
\title{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object containing 
simulated counts.}
\format{
An object of class 
\code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}
\source{
https://www.rhondabacher.com/scaffold-vignette.pdf
}
\usage{
data(sim_counts)
}
\value{
Small simulated trajectory dataset.
}
\description{
Data simulated using the \code{scaffold} R package for 50 dynamic and 50 
static genes across 1200 cells from 3 subjects.
}
\keyword{datasets}
