% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_pseudotime}
\alias{sim_pseudotime}
\title{A data.frame containing ground-truth pseudotime.}
\format{
An object of class \code{data.frame} with 1200 rows and one variable:
\itemize{
 \item PT: the true pseudotime (0.0025--1)
}
}
\usage{
data(sim_pseudotime)
}
\value{
Small simulated trajectory pseudotime.
}
\description{
The true ordering of the 1200 cells contained in \code{sim_counts}.
}
\keyword{datasets}
