% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothedCountsMatrix.R
\name{smoothedCountsMatrix}
\alias{smoothedCountsMatrix}
\title{Generate a smoothed matrix of gene expression using \code{scLANE} models.}
\usage{
smoothedCountsMatrix(
  test.dyn.res = NULL,
  genes = NULL,
  parallel.exec = TRUE,
  n.cores = 2
)
}
\arguments{
\item{test.dyn.res}{The list returned by \code{\link{testDynamic}} - no extra processing required. Defaults to NULL.}

\item{genes}{(Optional) A character vector of genes with which to subset the results. Defaults to NULL.}

\item{parallel.exec}{Should \code{furrr} be used to speed up execution? Defaults to TRUE.}

\item{n.cores}{If parallel execution is desired, how many cores should be utilized? Defaults to 2.}
}
\value{
A list of matrices of smoothed counts, with each element of the list being a single pseudotime lineage.
}
\description{
This function takes as input the output from \code{\link{testDynamic}} and returns the fitted values from each model in a wide format, with one column per-gene and one row-per cell. This matrix can be use as input to cell or gene clustering and / or visualizations such as heatmaps.
}
\examples{
\dontrun{
smoothedCountsMatrix(gene_stats)
smoothedCountsMatrix(gene_stats,
                     genes = c("AURKA", "KRT19", "EPCAM", "GATA6"),
                     parallel.exec = TRUE,
                     n.cores = 2)
}
}
\seealso{
\code{\link{testDynamic}}
}
\author{
Jack Leary
}
