% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripGLM.R
\name{stripGLM}
\alias{stripGLM}
\title{Make GLM objects much smaller.}
\usage{
stripGLM(glm.obj = NULL)
}
\arguments{
\item{glm.obj}{An object of class GLM from which you'd like to strip out unnecessary components.}
}
\value{
A slimmed-down \code{glm} object.
}
\description{
This function removes a \emph{lot} of components from the default GLM object in order to make it take up less memory. It does however retain enough pieces for \code{predict()} to still work. No promises beyond that.
}
\examples{
stripGLM(marge_model)
}
