% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModel.R
\name{summarizeModel}
\alias{summarizeModel}
\title{Represent a \code{marge} model as a series of piecewise equations.}
\usage{
summarizeModel(marge.model = NULL, pt = NULL)
}
\arguments{
\item{marge.model}{The fitted model output from \code{\link{marge2}} (this function is internally called by \code{\link{testDynamic}}). Defaults to NULL.}

\item{pseudotime_df}{The predictor matrix of pseudotime. Defaults to NULL.}
}
\value{
A data.frame of the model coefficients, cutpoint intervals, and formatted equations.
}
\description{
This function summarizes the model for each gene and allows for quantiative interpretation of fitted gene dynamics..
}
\examples{
\dontrun{
summarizeModel(marge.model = marge_mod)
}
}
\author{
Jack Leary & Rhonda Bacher
}
