% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp1.R
\name{tp1}
\alias{tp1}
\title{Truncated p-th power function (positive part).}
\usage{
tp1(x, t, p = 1)
}
\arguments{
\item{x}{: a vector of predictor variable values.}

\item{t}{: a specified knot value.}

\item{p}{: the pth degree of the polynomial considered.}
}
\value{
\code{tp1} returns a vector of values that have been transformed using a truncated p-th power function (positive part) for a specified knot value.
}
\description{
Truncated p-th power function (positive part).
}
\examples{
data(leptrine)
dat1 <- leptrine[[1]]
X_pred <- dat1[, 1]  # One predictor used.
tp1(X_pred, 1)  # Knot value set at x = 1.
}
\references{
Friedman, J. (1991). Multivariate adaptive regression splines. \emph{The Annals of Statistics}, \strong{19}, 1--67.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, 245--253.
}
\seealso{
\code{\link{tp2}}
}
\author{
Jakub Stoklosa and David I. Warton
}
