% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedGenes.R
\name{embedGenes}
\alias{embedGenes}
\title{Generate a table of fitted values and celltype metadata for genes of interest.}
\usage{
embedGenes(
  smoothed.counts = NULL,
  genes = NULL,
  pc.embed = 30,
  pcs.return = 2,
  cluster.genes = TRUE,
  gene.meta.data = NULL,
  k.param = 20,
  resolution.param = NULL,
  random.seed = 312
)
}
\arguments{
\item{smoothed.counts}{The output from \code{\link{smoothedCountsMatrix}}. Defaults to NULL.}

\item{genes}{A character vector of genes to embed. If not specified, all genes in \code{smoothed.counts} are used. Defaults to NULL.}

\item{pc.embed}{(Optional) How many PCs should be used to cluster the genes and run UMAP? Defaults to 30.}

\item{pcs.return}{(Optional) How many principal components should be included in the output? Defaults to 2.}

\item{cluster.genes}{(Optional) Should genes be clustered in PCA space using the Leiden algorithm? Defaults to TRUE.}

\item{gene.meta.data}{(Optional) A data.frame of metadata values for each gene (HVG status, Ensembl ID, gene biotype, etc.) that will be included in the result table. Defaults to NULL.}

\item{k.param}{(Optional) The value of nearest-neighbors used in creating the SNN graph prior to clustering & in running UMAP. Defaults to 20.}

\item{resolution.param}{(Optional) The value of the resolution parameter for the Leiden algorithm. If unspecified, silhouette scoring is used to select an optimal value. Defaults to NULL.}

\item{random.seed}{(Optional) The random seed used to control stochasticity in the clustering algorithm. Defaults to 312.}
}
\value{
A data.frame containing embedding coordinates, cluster IDs, and metadata for each gene.
}
\description{
Embed genes in dimension-reduced space given a smoothed counts matrix.
}
\examples{
\dontrun{
embedGenes(smoothed_counts$Lineage_A,
           pcs.return = 3,
           cluster.genes = TRUE)
}
}
\author{
Jack Leary
}
