% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBreakpoints.R
\name{extractBreakpoints}
\alias{extractBreakpoints}
\title{Identify breakpoints in a \code{marge} model.}
\usage{
extractBreakpoints(model = NULL, sort = FALSE)
}
\arguments{
\item{model}{The \code{marge} model to analyze. Defaults to NULL.}

\item{sort}{Should the breakpoints be sorted into ascending order? Defaults to FALSE.}
}
\value{
A numeric vector of breakpoints.
}
\description{
Extracts the breakpoints from a fitted \code{marge} model. Note - this function relies on the name of the pseudotime variable not having any numeric characters in it e.g., "pseudotime" would be fine but "pseudotime1" would not.
}
\examples{
extractBreakpoints(model = marge_mod)
}
