% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBreakpoints.R
\name{extractBreakpoints}
\alias{extractBreakpoints}
\title{Identify breakpoints}
\usage{
extractBreakpoints(model = NULL)
}
\arguments{
\item{model}{The \code{marge} model to analyze. Defaults to NULL.}
}
\value{
A numeric vector of breakpoints.
}
\description{
Extracts the breakpoints from a fitted \code{marge} model.
}
\examples{
extractBreakpoints(model = marge_mod)
}
