% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResultsDE.R
\name{getResultsDE}
\alias{getResultsDE}
\title{Tidy the results of \code{\link{testDynamic}}.}
\usage{
getResultsDE(
  test.dyn.results = NULL,
  p.adj.method = "bonferroni",
  fdr.cutoff = 0.01
)
}
\arguments{
\item{test.dyn.results}{The nested list returned by \code{\link{testDynamic}}. Defaults to NULL.}

\item{p.adj.method}{(Optional) The method used to adjust \emph{p}-values for MHT. Defaults to "bonferroni".}

\item{fdr.cutoff}{The FDR threshold for determining statistical significance. Defaults to 0.01.}
}
\value{
A data.frame containing differential expression results & modeling statistics for each gene.
}
\description{
This function turns the nested list differential expression results of \code{\link{testDynamic}} and turns them into a tidy \code{data.frame}.
}
\examples{
\dontrun{GetResultsDE(test.dyn.results = gene_stats)}
#' \dontrun{GetResultsDE(test.dyn.results = gene_stats, p.adj.method = "BH")}
}
\seealso{
\code{\link{testDynamic}} \code{\link{p.adjust}}
}
\author{
Jack Leary
}
