% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResultsDE.R
\name{GetResultsDE}
\alias{GetResultsDE}
\title{A function to tidy the results of \code{scLANE()}.}
\usage{
GetResultsDE(scLANE.results = NULL, p.adj.method = "BH")
}
\arguments{
\item{scLANE.results}{The nested list returned by \code{scLANE()}. Defaults to NULL.}

\item{p.adj.method}{(Optional) The method used to adjust \emph{p}-values for MHT. Defaults to "BH", for Benjamini-Hochberg.}
}
\value{
A data.frame containing differential expression results & modeling statistics for each gene.
}
\description{
This function turns the nested list differential expression results of \code{scLANE()} and turns them into a tidy data.frame, minus the \code{ggplot} objects.
}
\examples{
GetResultsDE(scLANE.results = scLANE_res_list)
}
\seealso{
\code{\link{scLANE}} \code{\link{p.adjust}}
}
