% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marge2.R
\name{marge2}
\alias{marge2}
\title{Fit \code{MARGE} models of single cell counts.}
\usage{
marge2(
  X_pred = NULL,
  Y = NULL,
  M = 5,
  is.gee = FALSE,
  id.vec = NULL,
  cor.structure = "independence",
  approx.knot = TRUE,
  pen = 2,
  tols_score = 1e-05,
  minspan = NULL,
  return.basis = FALSE,
  return.wic = FALSE,
  return.GCV = FALSE
)
}
\arguments{
\item{X_pred}{A matrix of the predictor variables. Note that this matrix should include a column of 1's for the intercept term.}

\item{Y}{The response variable. A vector of length n by N.}

\item{M}{A set threshold for the number of basis functions to be used. Defaults to 5.}

\item{is.gee}{Should the \code{\link[geeM]{geem}} package be used to fit a negative binomial GEE model? Defaults to FALSE.}

\item{id.vec}{If \code{is.gee = TRUE}, must be a vector of ID values for the observations. Data must be sorted such that the subjects are in order! Defaults to NULL.}

\item{cor.structure}{If \code{is.gee = TRUE}, must be a string specifying the desired correlation structure for the NB GEE. Defaults to "independence".}

\item{approx.knot}{Should the space of possibke knots be reduce in order to speed up computation? This has little effect on the final fit, but can improve computation time significantly. Defaults to TRUE.}

\item{pen}{(Optional) A set penalty used for the GCV (note: MARGE doesn't actually use this). The default is 2.}

\item{tols_score}{(Optional) The set tolerance for monitoring the convergence for the difference in score statistics between the parent and candidate model (this is the lack-of-fit criterion used for MARGE). The default is 0.00001}

\item{minspan}{A set minimum span value. The default is \code{minspan = NULL}.}

\item{return.basis}{(Optional) Whether the basis model matrix (denoted \code{B_final}) should be returned as part of the \code{marge} model object. Defaults to FALSE since it makes the model object much larger than necessary.}

\item{return.wic}{(Optional) Whether the WIC matrix and final WIC value should be returned as part of the \code{marge} model object. Defaults to FALSE.}

\item{return.GCV}{(Optional) Whether the final GCV value should be returned as part of the \code{marge} model object. Defaults to FALSE.}
}
\value{
An object of class \code{marge} containing the fitted model & other optional quantities of interest (basis function matrix, GCV, etc.).
}
\description{
MARS fitting function for generalized linear models (GLMs).
}
\examples{
\dontrun{
  marge2(X_pred = pseudotime_df,
         Y = expr_vec,
         M = 3)
  marge2(X_pred = pseudotime_df,
         Y = expr_vec,
         is.gee = TRUE,
         id.vec = subject_vec,
         cor.structure = "exchangeable")
}
}
\references{
Friedman, J. (1991). Multivariate adaptive regression splines. \emph{The Annals of Statistics}, \strong{19}, 1--67.

Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, \strong{70}, 110--120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, 245--253.
}
\seealso{
\code{\link{mars_ls}}

\code{\link{backward_sel_WIC}}

\code{\link{testDynamic}}

\code{\link[MASS]{glm.nb}}

\code{\link[geeM]{geem}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
