% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbGAM.R
\name{nbGAM}
\alias{nbGAM}
\title{Fit a negative-binomial GAM.}
\usage{
nbGAM(expr = NULL, pt = NULL, id.vec = NULL)
}
\arguments{
\item{expr}{A vector of integer counts. Defaults to NULL.}

\item{pt}{A dataframe or vector of pesudotime values. Defaults to NULL.}

\item{id.vec}{(Optional) A vector of subject IDs to be used in creating random intercepts in the GAM. Useful for comparing GAMs to GLMMs. Defaults to NULL.}
}
\value{
An object of class \code{gamlss}
}
\description{
Fits a negative-binomial family GAM using a b-spline on pseudotime.
}
\examples{
\dontrun{
nbGAM(expr_vec, pt = pt_df)
nbGAM(expr_vec,
      pt = pt_df,
      id.vec = subject_ids)
}
}
\seealso{
\code{\link[gamlss]{gamlss}}

\code{\link[gamlss]{random}}
}
\author{
Jack Leary
}
