% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusteredGenes.R
\name{plotClusteredGenes}
\alias{plotClusteredGenes}
\title{Generate tidy results from \code{\link{clusterGenes}} to use in plotting.}
\usage{
plotClusteredGenes(
  test.dyn.res = NULL,
  gene.clusters = NULL,
  pt = NULL,
  parallel.exec = TRUE,
  n.cores = 2
)
}
\arguments{
\item{test.dyn.res}{The list returned by \code{\link{testDynamic}} - no extra processing required. Defaults to NULL.}

\item{gene.clusters}{The data.frame returned by \code{\link{clusterGenes}}. Defaults to NULL.}

\item{pt}{A data.frame containing the pseudotime or latent time estimates for each cell. Defaults to NULL.}

\item{parallel.exec}{Should \code{furrr} be used to speed up execution? Defaults to TRUE.}

\item{n.cores}{If parallel execution is desired, how many cores should be utilized? Defaults to 2.}
}
\value{
A \code{data.frame} with ready-to-plot tidy data. Includes columns for gene name, pseudotime lineage, cell name, fitted values on link & response scale, pseudotime, & gene cluster.
}
\description{
Generate a table of per-lineage, per-cluster fitted values from \code{scLANE} to be used in visualizations.
}
\examples{
\dontrun{
plotClusteredGenes(test.dyn.res = gene_stats,
                   gene.clusters = gene_clusters,
                   pt = pt_df) \%>\%
  ggplot(aes(x = PT, y = FITTED, color = CLUSTER, group = GENE) +
  geom_line() +
  facet_wrap(~LINEAGE))
}
}
\seealso{
\code{\link{clusterGenes}}
}
\author{
Jack Leary
}
