% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusteredGenes.R
\name{plotClusteredGenes}
\alias{plotClusteredGenes}
\title{Plot the results from \code{clusterGenes}.}
\usage{
plotClusteredGenes(
  test.dyn.results = NULL,
  gene.clusters = NULL,
  pt = NULL,
  parallel.exec = TRUE,
  n.cores = 2
)
}
\arguments{
\item{test.dyn.results}{The list returned by \code{\link{testDynamic}} - no extra processing required. Defaults to NULL.}

\item{gene.clusters}{The data.frame returned by \code{\link{clusterGenes}}. Defaults to NULL.}

\item{pt}{A data.frame containing a single column - the pseudotime or latent time estimates for each cell. Defaults to NULL.}

\item{parallel.exec}{Should \code{furrr} be used to speed up execution at the cost of increased memory usage? Defaults to TRUE.}

\item{n.cores}{If parallel execution is desired, how many cores should be utilized? Defaults to 2.}
}
\value{

}
\description{
Plot per-lineage, per-cluster fitted values from \code{scLANE}.
}
\examples{
\dontrun{}
}
\seealso{

}
\author{
Jack Leary
}
