% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.scLANE}
\alias{print.summary.scLANE}
\title{Print method for summary.scLANE objects.}
\usage{
\method{print}{summary.scLANE}(x)
}
\arguments{
\item{x}{An object of class summary.scLANE.}
}
\value{
A printed summary of overall scLANE results
}
\description{
Print method for summary.scLANE objects.
}
\author{
Jack R. Leary
}
