% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLANE.R
\name{scLANE}
\alias{scLANE}
\title{Identify differentially expressed genes over pseudotime.}
\usage{
scLANE(
  expr = NULL,
  pt = NULL,
  genes = NULL,
  n.cores = NULL,
  n.boot = 200,
  track.time = TRUE
)
}
\arguments{
\item{expr}{A matrix of integer-valued counts. Defaults to NULL.}

\item{pt}{A data.frame containing a single column - the pseudotime or latent time estimates for each cell. Defaults to NULL.}

\item{genes}{A character vector of genes to model. Defaults to NULL.}

\item{n.cores}{The number of CPUs to use when fitting models in parallel. Defaults to \code{parallel::detectCores() - 1}.}

\item{n.boot}{The number of bootstrap samples to use in calculating the \code{marge} model's \emph{p}-value. Defaults to 200.}

\item{track.time}{A boolean indicating whether the time it takes to execute \code{scLANE()} should be tracked & printed at the end. Defaults to TRUE.}
}
\value{
A list of lists; each sublist contains a gene name, default \code{marge} vs. null model test results & model statistics, bootstrapped \code{marge} vs. null model test results & model statistics, and a \code{ggplot} of the models' fitted values. Use \code{\link{getResultsDE}} to tidy the results.
}
\description{
This function uses adaptive NB GLM spline models as implemented in the \code{marge} package to identify non-linear patterns in gene expression across pseudotime.
}
\examples{
scLANE(expr = raw_counts, pt = pseudotime_df, genes = my_gene_vec)
}
