% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fun_glm.R
\name{score_fun_glm}
\alias{score_fun_glm}
\title{Given estimates from the null model fit and the design matrix for alternative model, find the score statistic (this is used for GLMs only).}
\usage{
score_fun_glm(
  Y = NULL,
  VS.est_list = NULL,
  A_list = NULL,
  B1_list = NULL,
  mu.est = NULL,
  V.est = NULL,
  B1 = NULL,
  XA = NULL
)
}
\arguments{
\item{Y}{: the response variable.}

\item{VS.est_list}{: a product of matrices.}

\item{A_list}{: a product of matrices.}

\item{B1_list}{: a product of matrices.}

\item{mu.est}{: estimates of the fitted mean under the null model.}

\item{V.est}{: estimates of the fitted variance under the null model.}

\item{B1}{: model matrix under the null model.}

\item{XA}{: model matrix under the alternative model.}
}
\value{
\code{score_fun_glm} returns a calculated score statistic for the null and alternative model when fitting a GLM.
}
\description{
Given estimates from the null model fit and the design matrix for alternative model, find the score statistic (this is used for GLMs only).
}
\references{
Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, \strong{70}, 110--120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, 245--253.
}
\author{
Jakub Stoklosa and David I. Warton.
}
