% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreTestGEE.R
\name{scoreTestGEE}
\alias{scoreTestGEE}
\title{Use a Lagrange multiplier (score) test to compare nested GEE models.}
\usage{
scoreTestGEE(
  mod.1 = NULL,
  mod.0 = NULL,
  alt.df = NULL,
  null.df = NULL,
  id.vec = NULL,
  cor.structure = "ar1"
)
}
\arguments{
\item{mod.1}{The model under the alternative hypothesis. Must be of class \code{geem}. Defaults to NULL.}

\item{mod.0}{The model under the null hypothesis. Must be of class \code{geem}. Defaults to NULL.}

\item{alt.df}{The dataframe used to fit the alternative model. Defaults to NULL.}

\item{null.df}{The dataframe used to fit the null model. Defaults to NULL.}

\item{id.vec}{A vector of subject IDs to use as input to \code{\link{marge2}}. Defaults to NULL.}

\item{cor.structure}{A string specifying the working correlation structure used to fit each model. Must be one of "ar1", "independence", or "exchangeable". Defaults to "ar1".}
}
\value{
A list containing the Score test statistic, a \emph{p}-value, and the degrees of freedom used in the test.
}
\description{
Performs a basic Lagrange multiplier test to determine whether an alternate model is significantly better than a nested null model. This is the GEE equivalent (kind of) of \code{\link{modelLRT}}. Be careful with small sample sizes.
}
\details{
\itemize{
\item Calculating the test statistic involves taking the inverse of the variance-covariance matrix of the coefficients. Ideally this would be done using the "true" inverse with something like \code{\link{solve}}, \code{\link{qr.solve}}, or \code{\link{chol2inv}}, but in practice this can cause issues when the variance-covariance matrix is near-singular. With this in mind, we use the Moore-Penrose pseudoinverse as implemented in \code{\link[MASS]{ginv}} instead, which leads to more stable results.
\item The \emph{p}-value is calculated using an asymptotic \eqn{\Chi^2} distribution, with the degrees of freedom equal to the number of non-intercept coefficients in the alternative model.
}
}
\seealso{
\code{\link[geeM]{geem}}

\code{\link{waldTestGEE}}

\code{\link{modelLRT}}
}
\author{
Jack R. Leary
}
