% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModel.R
\name{summarizeModel}
\alias{summarizeModel}
\title{Represent a \code{marge} model as a series of piecewise equations.}
\usage{
summarizeModel(marge.model = NULL)
}
\arguments{
\item{marge.model}{The fitted model output from \code{\link{marge2}} (this function is internally called by \code{\link{testDynamic}}). Defaults to NULL.}
}
\value{
A data.frame of the model coefficients, cutpoint intervals, and formatted equations.
}
\description{
This function converts a \code{marge} model from using truncated power basis functions to a more interpretable series of piecewise linear equations over a set of empirically chosen cutpoint intervals. Note that the function assumes that pseudotime has been normalized to exist over \eqn{[0, 1]}.
}
\examples{
\dontrun{
summarizeModel(marge.model = marge_mod)
}
}
\author{
Jack Leary
}
