% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.scLANE}
\alias{summary.scLANE}
\title{Summary method for scLANE objects.}
\usage{
\method{summary}{scLANE}(object, ...)
}
\arguments{
\item{test.dyn.res}{The nested list returned by \code{\link{testDynamic}}.
Defaults to NULL.}
}
\value{
A summary list with aggregated statistics concerning the trajectory
DE tests from \code{scLANE}.
}
\description{
Summary method for scLANE objects.
}
\examples{
data(scLANE_models)
summary(scLANE_models)
}
\author{
Jack R. Leary
}
