% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.scLANE}
\alias{summary.scLANE}
\title{Summary method for scLANE objects.}
\usage{
\method{summary}{scLANE}(test.dyn.res = NULL, p.adj.method = "fdr", fdr.cutoff = 0.01)
}
\arguments{
\item{test.dyn.res}{The nested list returned by \code{\link{testDynamic}}. Defaults to NULL.}

\item{p.adj.method}{(Optional) The method used to adjust \emph{p}-values for multiple hypothesis testing. Defaults to "fdr".}

\item{fdr.cutoff}{(Optional) The FDR threshold for determining statistical significance. Defaults to 0.01.}
}
\value{
A summary list with aggregated statistics concerning the trajectory DE tests from \code{scLANE}.
}
\description{
Summary method for scLANE objects.
}
\examples{
data(scLANE_models)
summary(scLANE_models)
}
\author{
Jack R. Leary
}
