% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSlope.R
\name{testSlope}
\alias{testSlope}
\title{Test whether a gene's expression is changing over a window of pseudotime.}
\usage{
testSlope(
  test.dyn.results = NULL,
  p.adj.method = "bonferroni",
  fdr.cutoff = 0.01
)
}
\arguments{
\item{test.dyn.results}{The list returned by \code{\link{testDynamic}} - no extra processing required. Defaults to NULL.}

\item{p.adj.method}{The method used to adjust the \emph{p}-values for each slope. Defaults to "bonferroni".}

\item{fdr.cutoff}{The FDR threshold for determining statistical significance. Defaults to 0.01.}
}
\value{
A dataframe containing the genes, breakpoints, and slope \emph{p}-values from each model.
}
\description{
This function tests whether the slope of a gene's \eqn{\beta} coefficient(s) for pseudotime has a significant effect on expression within a pseudotime window. In short, it tells us whether a gene's expression is changing over an interval or not.
}
\examples{
\dontrun{
testSlope(test.dyn.results = gene_stats)
testSlope(test.dyn.results = gene_stats, method = "BH")
}
}
\seealso{
\code{\link{testDynamic}}

\code{\link[stats]{p.adjust}}
}
\author{
Jack Leary
}
