% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSlope.R
\name{testSlope}
\alias{testSlope}
\title{Test whether a gene is dynamic over a pseudotime interval.}
\usage{
testSlope(test.dyn.results = NULL, p.adj.method = "holm", fdr.cutoff = 0.01)
}
\arguments{
\item{test.dyn.results}{The list returned by \code{\link{testDynamic}} - no extra processing required. Defaults to NULL.}

\item{p.adj.method}{The method used to adjust the \emph{p}-values for each slope. Defaults to "holm".}

\item{fdr.cutoff}{The FDR threshold for determining statistical significance. Defaults to 0.01.}
}
\value{
A dataframe containing the genes, breakpoints, and slope \emph{p}-values from each model.
}
\description{
This function tests whether each gene's estimated \eqn{\beta} for pseudotime differs significantly from 0 over each empirically estimated sets of knots / pseudotime interval using a Wald test.
}
\examples{
\dontrun{
testSlope(test.dyn.results = gene_stats)
testSlope(test.dyn.results = gene_stats, method = "BH", fdr.cutoff = 0.05)
}
}
\seealso{
\code{\link{testDynamic}}

\code{\link[stats]{p.adjust}}
}
\author{
Jack Leary
}
