% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldTestGEE.R
\name{waldTestGEE}
\alias{waldTestGEE}
\title{Use a Wald test to compare nested GEE models.}
\usage{
waldTestGEE(mod.1 = NULL, mod.0 = NULL)
}
\arguments{
\item{mod.1}{The alternate model. Must be of class \code{geem}. Defaults to NULL.}

\item{mod.0}{The null model. Must be of class \code{geem}. Defaults to NULL.}
}
\value{
A list containing the Wald test statistic, a \emph{p}-value, and the degrees of freedom used in the test.
}
\description{
Performs a basic Wald test to determine whether an alternate model is significantly better than a nested null model. This is the GEE equivalent (kind of) of \code{\link{modelLRT}}. Be careful with small sample sizes.
}
\examples{
\dontrun{waldTestGEE(mod.1 = full_model, mod.0 = null_model)}
}
\seealso{
\code{\link[multgee]{waldts}}

\code{\link[geeM]{geem}}

\code{\link{modelLRT}}
}
\author{
Jack Leary
}
